Creo módulo de Book para poderlo utilizar como dependencia
