--- examples/book.py
@@ -1,6 +1,0 @@
-import Book
-
-if __name__ == "__main__":
-    libro = Book()
-    libro.load_file("test.bg")
-    libro.init()
